refactor: 소모임 서비스 디렉토리 이동
- service 폴더 생성 후 이동
- #26
